+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use `proxyInterceptor` hooks in `proxy` middleware to apply custom logic on the request and responses instead of patching them directly ([PORTAL-1146](https://issue.inventage.com/browse/PORTAL-1146)).
 - Patched issuer URL of JWT returned by the OIDC flow to the publicly facing URL ([PORTAL-1146](https://issue.inventage.com/browse/PORTAL-1146)).
 - Inherit from public class `JWTAuthHandler` (and its ecosystem) to apply custom claim checks instead of copying the whole Vert.x JWT ecosystem ([PORTAL-1146](https://issue.inventage.com/browse/PORTAL-1146)).
+- `options` in middlware configurations should be optional, if no options are needed ([PORTAL-1240](https://issue.inventage.com/browse/PORTAL-1240)).
 
 ### Removed